Add a troubleshooting item for setting the right locale
If the system uses a locale with the comma as decimal separator (in my case German), the compile goes fine but when launching, OpenGL errors will pop up for all planets and only the atmospheres are visible.

This behavior has been described in [issue 1442 on the OpenSpace repo](https://github.com/OpenSpace/OpenSpace/issues/1442) but since the issue is quite hard to find and also doesn't result in the same error messages as with v16.0, I thought it would be good to point this out to everyone compiling on Ubuntu. Maybe it would make sense to also copy this to the compile guide for fedora?